Added JSON-encoded user color bars to tile request URL; extracted new module @/model/userColorBars
