+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## Changes in version 1.2.0 (in development)
 
+* Users can now define their own color bars.
+  This feature requires xcube server >= 1.6.
+
 * Introduced overlay layers that can be selected in the settings.
 
 * Users can now define their own base maps and overlay layers.
@@ -40,10 +43,10 @@
 TODO:
 
 * User color bars:
-  - Implement user color bars in xcube server and adjust tile API calls
-    in viewer
   - Harmonize styling see useItemStyles.ts, search for COLOR_BAR_ITEM_GAP
   - Support categorical user color bars
+  - ~Implement user color bars in xcube server and adjust tile API calls
+    in viewer~
   - ~Use RGBA instead of just RGB~
   - ~Fix code duplication, search for COLOR_BAR_ITEM_BOX_MARGIN~
   - ~Fix popover toolbar for editing and removing a color bar~
